backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-Cors>=3.0
python-dotenv>=0.19
Werkzeug>=2.0
